ml/requirements.txt
torch>=2.2
torchvision>=0.17
numpy>=1.24
Pillow>=10.0
scikit-learn>=1.3
tqdm>=4.66
matplotlib>=3.8
onnx>=1.15
onnxruntime>=1.17
facenet-pytorch>=2.5.3
huggingface_hub>=0.20
kagglehub>=0.2
requests>=2.31
certifi>=2024.0
pyarrow>=14.0
